requirements.txt
google-adk
py-trello
datetime
dotenv